+++ lib/src/bloc/completable_bloc.dart
@@ -1,10 +1,12 @@
 import 'package:bloc_process/src/bloc/interfaces/process_bloc_event.dart';
 import 'package:bloc_process/src/bloc/interfaces/process_bloc_state.dart';
+import 'package:bloc_process/src/bloc/middleware_bloc.dart';
 import 'package:flutter/cupertino.dart';
-import 'package:flutter_bloc/flutter_bloc.dart';
 
-abstract class CompletableBloc<TEvent extends ProcessBlocEvent,
-    TState extends ProcessBlocState, TReturn> extends Bloc<TEvent, TState> {
+abstract class CompletableBloc<
+    TEvent extends ProcessBlocEvent,
+    TState extends ProcessBlocState,
+    TReturn> extends MiddlewareBloc<TEvent, TState> {
   @protected
   late final void Function(TReturn value) completeCallback;