+++ lib/src/middleware/error_middleware.dart
@@ -1,0 +1,6 @@
+import '../../bloc_process.dart';
+
+abstract interface class ErrorMiddleware<TError extends ErrorEvent>
+    implements Middleware {
+  Future<void> onError(TError error);
+}