+++ lib/src/middleware/event_middleware.dart
@@ -1,0 +1,6 @@
+import 'package:bloc_process/bloc_process.dart';
+
+abstract interface class EventMiddleware<TEvent extends ProcessBlocEvent>
+    implements Middleware {
+  Future<void> onEvent(TEvent event);
+}