+++ lib/src/middleware/state_middleware.dart
@@ -1,0 +1,6 @@
+import 'package:bloc_process/bloc_process.dart';
+
+abstract interface class StateMiddleware<TState extends ProcessBlocState>
+    implements Middleware {
+  Future<void> onStateChanged(TState state);
+}