+++ test/src/process/middleware/state_test_middleware.dart
@@ -1,0 +1,11 @@
+import 'package:bloc_process/src/bloc/interfaces/process_bloc_state.dart';
+import 'package:bloc_process/src/middleware/state_middleware.dart';
+import 'package:mockito/annotations.dart';
+
+@GenerateNiceMocks([MockSpec<StateTestMiddleware>()])
+class StateTestMiddleware implements StateMiddleware {
+  @override
+  Future<void> onStateChanged(ProcessBlocState state) {
+    return Future.value();
+  }
+}