+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.2.0
+effect stream support
+middleware support
+
+## 0.1.1
+multiple fixes for MultiChannelProcessBloc
+
 ## 0.1.0
 added support for navigation configuration